+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPW92EGYRGPWN2FX3CSX50DY"
+        "01KPWHZ3JF05BYMVCE69A4QHFY"
     )
     r"""The unique identifier of the data source"""