+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPW92EH106CN21DTK8XE8Z0K"
+        "01KPWHZ3JGTAHENKVKJM6AJ5YM"
     )
     r"""The unique identifier of the data source"""