+++ packages/orq-rc/RELEASES.md
@@ -7269,3 +7269,13 @@ Based on:
 - [python v4.8.0-rc.50] packages/orq-rc
 ### Releases
 - [PyPI v4.8.0-rc.50] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.50 - packages/orq-rc
+
+## 2026-04-23 06:58:35
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.8.0-rc.51] packages/orq-rc
+### Releases
+- [PyPI v4.8.0-rc.51] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.51 - packages/orq-rc